DomotikBox/pyproject.toml
[tool.poetry]
name = "domotikbox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
paho-mqtt = "^2.1.0"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
